revert one change

+++ auto_round/data_type/int.py
@@ -49,7 +49,7 @@ def quant_tensor_asym(weight, bits=4, v=0, min_scale=1.0, max_scale=1.0, scale_d
         wmax = wmax_tmp
     scale = ((wmax - wmin) / maxq).to(scale_dtype)
     scale = torch.clamp(scale, min=q_scale_thresh)
-    zp = round_ste(-wmin / scale).to(weight.dtype)  # pylint: disable=E1130
+    zp = round_ste(-wmin / scale) # pylint: disable=E1130
     scale = scale.unsqueeze(dim=-1)
     zp = zp.unsqueeze(dim=-1)
     int_w = round_ste(weight / scale + v)
